% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daive_grid.R
\name{daive_grid}
\alias{daive_grid}
\title{DAIVE grid}
\usage{
daive_grid(
  data,
  outcomes,
  components,
  costs = "default",
  separator = "",
  scale = "0-1",
  weights = "default"
)
}
\arguments{
\item{data}{A list of draws objects.}

\item{outcomes}{A char (or vector of chars) with name(s) of the outcome(s).}

\item{components}{A char (or vector of chars) with name(s) of the component(s).}

\item{costs}{A numeric or ordered char of numerics of component delivery costs.}

\item{separator}{A char to separate component names in alternatives labels.}

\item{scale}{A function used to scale outcomes for comparison.}

\item{weights}{A numeric or ordered vector of numerics for outcome weights.}
}
\value{
A daive_grid object.
}
\description{
DAIVE grid
}
