% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{frontier_plot}
\alias{frontier_plot}
\title{Creates a scatterplot of expected alternative values with a frontier of
efficiency.}
\usage{
frontier_plot(g, frontier = NULL, outs = NULL, weights = NULL, static = FALSE)
}
\arguments{
\item{g}{A daive_grid object.}

\item{frontier}{A dataframe containing the values and costs associated with
interventions on the frontier of efficiency.}

\item{outs, weights}{Vectors of char values and numeric values containing the
outcome labels and weights respectively, used for generating an x axis label}

\item{static}{A boolean indicating whether the function should return an
interactive plotly object or a static ggplot object}
}
\description{
This function takes a value summary data frame and generates a plotly object
to visualize Decision Analysis for Intervention Value Efficiency. If no
frontier is provided, the method identifies a frontier of efficiency using
the outcome and cost values provided.
}
