% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{get_codes}
\alias{get_codes}
\title{Generate effect codes for each intervention version using component labels}
\usage{
get_codes(components)
}
\description{
Generate effect codes for each intervention version using component labels
}
