% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daive_grid.R
\name{get_draws}
\alias{get_draws}
\title{Get draws values.}
\usage{
get_draws(g)
}
\arguments{
\item{g}{A daive_grid object.}
}
\value{
A dataframe or list of draws contained in the input.
}
\description{
Get draws values.
}
