% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value.R
\name{get_ev}
\alias{get_ev}
\title{Generates a table of expected value for all possible combinations of
components.}
\usage{
get_ev(g)
}
\arguments{
\item{g}{A daive_grid object}
}
\value{
A daive_grid object
}
\description{
This function "("get expected value") updates a daive_grid object to include
a summary dataframe with estimated performance for each alternative and a
predicted_outcomes dataframe formatted to estimate predicted alternative
performance. The summary frame can be used to calculate a frontier and create
DAIVE plots.
}
