% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daive_grid.R
\name{get_predicted_outcomes}
\alias{get_predicted_outcomes}
\title{Get predicted outcomes.}
\usage{
get_predicted_outcomes(g)
}
\arguments{
\item{g}{A daive_grid object.}
}
\value{
A dataframe with the predicted values contained in the input.
}
\description{
Get predicted outcomes.
}
