% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daive_grid.R
\name{get_table}
\alias{get_table}
\title{Get expected value table.}
\usage{
get_table(g)
}
\arguments{
\item{g}{A daive_grid object.}
}
\value{
A dataframe with the expected values table contained in the input.
}
\description{
Get expected value table.
}
