% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{prepare_draws}
\alias{prepare_draws}
\title{Formats posterior draws to estimate alternative performance.}
\usage{
prepare_draws(draws, codes, components)
}
\description{
Formats posterior draws to estimate alternative performance.
}
