% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daive_grid.R
\name{print.daive_grid}
\alias{print.daive_grid}
\title{Print}
\usage{
\method{print}{daive_grid}(g)
}
\arguments{
\item{g}{A daive_grid object.}
}
\value{
Printed output with contents of the input.
}
\description{
Print
}
