% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daive_grid.R
\name{summary.daive_grid}
\alias{summary.daive_grid}
\title{Summary}
\usage{
\method{summary}{daive_grid}(g)
}
\arguments{
\item{g}{A daive_grid object.}
}
\value{
Printed output describing expected value and the frontier.
}
\description{
Summary
}
