% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{update_weights}
\alias{update_weights}
\title{Combines multiple outcomes using weighted sums.}
\usage{
update_weights(grid_object, weights)
}
\arguments{
\item{grid_object}{A daive_grid object}

\item{weights}{A vector of numeric values with weights for each outcome}
}
\value{
A vector containing the weighted sums of the specified columns for
each row of data.
}
\description{
This function takes a data frame, a vector of outcome labels, and a set of
numeric weights and generates a vector of combined outcome values using
simple additive weighting.
}
