% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{values_table}
\alias{values_table}
\title{Generates table of expected values (STILL IN DEVELOPMENT)}
\usage{
values_table(draws, outs, components, codes)
}
\description{
Generates table of expected values (STILL IN DEVELOPMENT)
}
