% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{DAIVE_plot}
\alias{DAIVE_plot}
\title{Creates a scatterplot of expected alternative values with a frontier of
efficiency.}
\usage{
DAIVE_plot(data, outcome, cost, frontier = NULL, outs = NULL, weights = NULL)
}
\description{
Creates a scatterplot of expected alternative values with a frontier of
efficiency.
}
