% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{DAIVE_plot}
\alias{DAIVE_plot}
\title{Creates a scatterplot of expected alternative values with a frontier of
efficiency.}
\usage{
DAIVE_plot(
  data,
  outcome,
  cost,
  frontier = NULL,
  outs = NULL,
  weights = NULL,
  static = FALSE
)
}
\arguments{
\item{data}{A data frame containing a summary of expected value for each
alternative}

\item{outcome}{A char value containing the column name of the outcome
variable (or value function output for multiple weighted outcomes)}

\item{cost}{A char value containing the colname of the cost variable}

\item{frontier}{A data frame containing the values and costs associated with
interventions on the frontier of efficiency}

\item{outs, weights}{Vectors of char values and numeric values containing the
outcome labels and weights respectively, used for generating an x axis label}

\item{static}{A boolean indicating whether the function should return an
interactive plotly object or a static ggplot object}
}
\value{
An interactive plotly object or static ggplot object (depending on
'static' parameter) containing a scatterplot with all alternatives, and a
frontier of efficiency denoting the value efficient alternatives.
}
\description{
This function takes a value summary data frame and generates a plotly object
to visualize Decision Analysis for Intervention Value Efficiency. If no
frontier is provided, the method identifies a frontier of efficiency using
the outcome and cost values provided.
}
