% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{DAIVE_plot}
\alias{DAIVE_plot}
\title{Generates scatterplot of expected values with frontier of efficiency}
\usage{
DAIVE_plot(data, outcome, cost, frontier = NULL)
}
\description{
Generates scatterplot of expected values with frontier of efficiency
}
