% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{get_codes}
\alias{get_codes}
\title{Generates effect codes for each intervention version using component labels.}
\usage{
get_codes(components)
}
\arguments{
\item{components}{A vector containing a char value with the names of each
component.}
}
\value{
A data frame with effect codes.
}
\description{
This function generates effect codes using a list of component names. Effect
codes will include an intercept, main effects, and interaction effects up to
the k-way interaction for k components.
}
\examples{
get_codes(c("A", "B", "C", "D"))
}
