% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{prepare_draws}
\alias{prepare_draws}
\title{Format tidybayes::spread_draws output for estimating alternative performance}
\usage{
prepare_draws(draws, codes, components)
}
\description{
Format tidybayes::spread_draws output for estimating alternative performance
}
