% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{weight_outcomes}
\alias{weight_outcomes}
\title{Combines multiple outcomes using weighted sums.}
\usage{
weight_outcomes(data, outcomes, weights)
}
\description{
Combines multiple outcomes using weighted sums.
}
